% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-keys.R
\name{draw_key_richtext}
\alias{draw_key_richtext}
\title{Key glyph for rich text}
\usage{
draw_key_richtext(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A \code{\link[gridtext:richtext_grob]{richtext_grob}} that represents
formatted text.
}
\description{
This is a function for rendering the key in the rich text style when a geom
needs to be displayed in a legend. It is designed to be provided to a layer's
\code{key_glyph} argument, either as a function or as \code{key_glyph = "richtext"}.
}
\examples{
ggplot(mtcars, aes(wt, mpg, label = rownames(mtcars))) +
  geom_richtext(aes(colour = factor(cyl)), key_glyph = "richtext")
}
