% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-markdown.R
\name{element_markdown}
\alias{element_markdown}
\title{Theme element that enables markdown text.}
\usage{
element_markdown(
  family = NULL,
  face = NULL,
  size = NULL,
  colour = NULL,
  fill = NULL,
  box.colour = NULL,
  linetype = NULL,
  linewidth = NULL,
  hjust = NULL,
  vjust = NULL,
  box.hjust = NULL,
  box.vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL,
  padding = NULL,
  r = NULL,
  color = NULL,
  box.color = NULL,
  rotate_margins = FALSE,
  debug = FALSE,
  inherit.blank = FALSE
)
}
\arguments{
\item{family}{Font family}

\item{face}{Font face}

\item{size}{Font size}

\item{colour, color}{Text color}

\item{hjust}{Horizontal justification}

\item{vjust}{Vertical justification}

\item{angle}{Angle (in degrees)}

\item{lineheight}{Line height}

\item{margin}{Margins around the text. See \code{\link[=margin]{margin()}} for
details.}

\item{rotate_margins}{Should margins get rotated in frame with rotated text?
If \code{TRUE}, the margins are applied relative to the text direction. If \code{FALSE},
the margins are applied relative to the plot direction, i.e., the top margin,
for example, is always placed above the text label, regardless of the direction
in which the text runs. The default is \code{FALSE}, which mimics the behavior of
\code{element_text()}.}

\item{debug}{Draw a debugging box around each label}

\item{inherit.blank}{See \code{\link[=margin]{margin()}} for details.}
}
\description{
Theme element that enables markdown text.
}
\seealso{
\code{\link[=richtext_grob]{richtext_grob()}}, \code{\link[=element_textbox]{element_textbox()}}
}
