% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-textbox.R
\name{element_textbox}
\alias{element_textbox}
\title{Theme element that enables markdown text in a box.}
\usage{
element_textbox(family = NULL, face = NULL, colour = NULL,
  size = NULL, hjust = NULL, vjust = NULL, angle = NULL,
  lineheight = NULL, color = NULL, margin = NULL, height = NULL,
  width = NULL, debug = FALSE, inherit.blank = FALSE)
}
\arguments{
\item{family}{Font family}

\item{face}{Font face}

\item{colour, color}{Text color}

\item{size}{Font size}

\item{hjust}{Horizontal justification}

\item{vjust}{Vertical justification}

\item{angle}{Angle (in degrees)}

\item{lineheight}{Line height}

\item{margin}{Margins around the text. See \code{\link[ggplot2:margin]{ggplot2::margin()}} for
details.}

\item{width, height}{Unit objects specifying the width and height
of the textbox, as in \code{\link[gridtext:textbox_grob]{gridtext::textbox_grob()}}.}

\item{debug}{Not implemented.}

\item{inherit.blank}{See \code{\link[ggplot2:margin]{ggplot2::margin()}} for details.}
}
\description{
Theme element that enables markdown text in a box. Doesn't work yet.
}
