% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-markdown.R
\name{element_markdown}
\alias{element_markdown}
\title{Theme element that enables markdown text.}
\usage{
element_markdown(family = NULL, face = NULL, colour = NULL,
  size = NULL, hjust = NULL, vjust = NULL, angle = NULL,
  lineheight = NULL, color = NULL, margin = NULL, debug = FALSE,
  inherit.blank = FALSE)
}
\arguments{
\item{family}{Font family}

\item{face}{Font face}

\item{colour, color}{Text color}

\item{size}{Font size}

\item{hjust}{Horizontal justification}

\item{vjust}{Vertical justification}

\item{angle}{Angle (in degrees)}

\item{lineheight}{Line height}

\item{margin}{Margins around the text. See \code{\link[ggplot2:margin]{ggplot2::margin()}} for
details.}

\item{debug}{Draw a debugging box around each label}

\item{inherit.blank}{See \code{\link[ggplot2:margin]{ggplot2::margin()}} for details.}
}
\description{
Theme element that enables markdown text.
}
