% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-textbox.R
\name{geom_textbox}
\alias{geom_textbox}
\title{Draw boxes containing text}
\usage{
geom_textbox(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., nudge_x = 0, nudge_y = 0,
  box.padding = unit(c(5.5, 5.5, 5.5, 5.5), "pt"), box.r = unit(5.5,
  "pt"), width = unit(2, "inch"), minwidth = NULL, maxwidth = NULL,
  height = NULL, minheight = NULL, maxheight = NULL, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\description{
Draw boxes of defined width and height containing word-wrapped text.
}
